backend/requirements.txt
absl-py==2.2.2
alembic==1.13.2
annotated-types==0.7.0
astunparse==1.6.3
av==14.3.0
beautifulsoup4==4.13.4
blinker==1.8.2
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
cryptography==44.0.2
ctranslate2 ==4.6.0
fanstatic==1.5
faster-whisper==1.1.1
filelock==3.18.0
Flask==3.0.3
flask-cors==5.0.1
Flask-Migrate==4.0.7
Flask-SQLAlche==3.1.1
flatbuffers==25.2.10
fsspec==2025.3.2
gast==0.6.0
google==3.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.167.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-cloud==0.34.0
google-cloud-core==2.4.3
google-cloud-translate==3.20.2
google-generativeai==0.8.5
google-pasta==0.2.0
googleapis-common-protos==1.70.0
greenlet==3.2.0
grpc-google-iam-v1==0.14.2
grpcio==1.72.0rc1
grpcio-status==1.71.0
h5py==3.13.0
hf-xet==1.0.3
httplib2==0.22.0
huggingface-hub==0.30.2
humanfriendly==10.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jiwer==3.1.0
js==1.0
keras==3.9.2
libclang==18.1.1
libretranslatepy==2.1.1
lxml==5.3.2
Mako==1.3.5
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ml_dtypes==0.5.1
mpmath==1.3.0
namex==0.0.8
networkx==3.4.2
numpy==2.1.3
onnxruntime==1.21.1
opt_einsum==3.4.0
optree==0.15.0
packaging==24.2
pandas==2.2.3
pip==25.0.1
pocketsphinx==5.0.4
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAudio==0.2.14
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
pydub==0.25.1
Pygments==2.19.1
PyMySQL==1.1.1
pyparsing==3.2.3
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
PyYAML==6.0.2
RapidFuzz==3.13.0
regex==2024.11.6
requests==2.32.3
rich==14.0.0
rsa==4.9.1
safetensors==0.5.3
sentencepiece==0.2.0
setuptools==78.1.0
six==1.17.0
sounddevice==0.5.1
soundfile==0.13.1
soupsieve==2.7
SpeechRecognition==3.14.2
SQLAlchemy==2.0.32
sympy==1.13.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
termcolor==3.0.1
tf_keras==2.19.0
tiktoken==0.9.0
tokenizers==0.21.1
torch==2.6.0
tqdm==4.67.1
transformers==4.51.3
translate==3.6.1
typing_extensions==4.12.2
typing-inspection==0.4.0
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
WebOb==1.8.9
Werkzeug==3.0.4
wheel==0.45.1
wrapt==1.17.2